+++ src/locale/multiple-select-ko-KR.js
@@ -15,6 +15,15 @@ $.fn.multipleSelect.locales['ko-KR'] = $.fn.multipleSelect.locales['ko'] = {
   },
   formatNoMatchesFound () {
     return '검색 결과가 없습니다.'
+  },
+  formatFilterAll () {
+    return '전체'
+  },
+  formatFilterSelected () {
+    return '선택됨'
+  },
+  formatFilterUnselected () {
+    return '선택 안됨'
   }
 }