+++ src/locale/multiple-select-pt-BR.js
@@ -15,6 +15,15 @@ $.fn.multipleSelect.locales['pt-BR'] = $.fn.multipleSelect.locales['pt'] = {
   },
   formatNoMatchesFound () {
     return 'Nenhum resultado encontrado'
+  },
+  formatFilterAll () {
+    return 'Todos'
+  },
+  formatFilterSelected () {
+    return 'Selecionados'
+  },
+  formatFilterUnselected () {
+    return 'Não selecionados'
   }
 }